requirements.txt
PyQt5>=5.15.7
Pillow>=10.0.0
PyInstaller>=6.0.0
